added apply_process flag
