+++ src/compas_timber/parts/beam.py
@@ -430,6 +430,8 @@ class Beam(Part):
         self.length = distance_point_point(ps, pe)
 
     def extension_to_plane(self, pln):
+
+
         """Returns the amount by which to extend the beam in each direction using metric units.
 
         TODO: verify this is true
@@ -455,12 +457,20 @@ class Beam(Part):
         de = 0.0
         if side == "start":
             tmin = min(x.keys())
-            if tmin < 0.0:
-                ds = tmin * self.length  # should be negative
+            ds = tmin * self.length  # should be negative
         elif side == "end":
             tmax = max(x.keys())
-            if tmax > 1.0:
-                de = (tmax - 1.0) * self.length
+            de = (tmax - 1.0) * self.length
+
+        # if side == "start":
+        #     tmin = min(x.keys())
+        #     if tmin < 0.0:
+        #         ds = tmin * self.length  # should be negative
+        # elif side == "end":
+        #     tmax = max(x.keys())
+        #     if tmax > 1.0:
+        #         de = (tmax - 1.0) * self.length
+
 
         return -ds, de
